+++ meteor/server/ActivationCache.ts
@@ -66,7 +66,7 @@ export class ActivationCache {
 	private _initialized: boolean = false
 	private _persistant: boolean = false
 
-	private _playlist: RundownPlaylist
+	private _playlist: RundownPlaylist | undefined
 	private _studio: Studio
 	private _showStyleBases: { [id: string]: InternalCache<ShowStyleBase> } = {}
 	private _showStyleVariants: { [id: string]: InternalCache<ShowStyleVariant> } = {}
@@ -108,7 +108,7 @@ export class ActivationCache {
 		this._persistant = false
 	}
 	async initialize(playlist: RundownPlaylist, rundownsInPlaylist: Rundown[]) {
-		if (this._initialized && playlist.activeInstanceId !== this._playlist.activeInstanceId) {
+		if (this._initialized && (!this._playlist || playlist.activeInstanceId !== this._playlist.activeInstanceId)) {
 			// activeInstanceId has changed, we should clear out the data because it might not be valid anymore
 			this._uninitialize()
 		}
@@ -176,6 +176,7 @@ export class ActivationCache {
 	}
 	getPlaylist(): RundownPlaylist {
 		if (!this._initialized) throw new Meteor.Error(`ActivationCache is not initialized`)
+		if (!this._playlist) throw new Meteor.Error(`ActivationCache is without playlist`)
 		return this._playlist
 	}
 	getStudio(): Studio {
@@ -206,7 +207,7 @@ export class ActivationCache {
 		if (!this._initialized) throw new Meteor.Error(`ActivationCache is not initialized`)
 		return this._getPeripheralDevices()
 	}
-	private async _getShowStyleBase(rundown: Rundown, supressError?: boolean): Promise<ShowStyleBase> {
+	private async _getShowStyleBase(rundown: Rundown): Promise<ShowStyleBase> {
 		if (!rundown.showStyleBaseId) throw new Meteor.Error(500, `Rundown.showStyleBaseId not set!`)
 		return this._getFromCache(this._showStyleBases, rundown.showStyleBaseId, '', async (id) => {
 			const showStyleBase = await asyncCollectionFindOne(ShowStyleBases, id)
@@ -260,12 +261,17 @@ export class ActivationCache {
 		)
 	}
 	private async _getPeripheralDevices(): Promise<PeripheralDevice[]> {
-		return this._getFromCache(this._peripheralDevices, this._playlist.studioId, '', async (id) => {
-			const devices = await asyncCollectionFindFetch(PeripheralDevices, {
-				studioId: id,
-			})
-			return devices
-		})
+		return this._getFromCache(
+			this._peripheralDevices,
+			this._playlist?.studioId ?? this._studio._id,
+			'',
+			async (id) => {
+				const devices = await asyncCollectionFindFetch(PeripheralDevices, {
+					studioId: id,
+				})
+				return devices
+			}
+		)
 	}
 	private _updateExpires() {
 		const TTL = 30 * 60 * 1000 // 30 minutes